Moving php installation to a pre start script
This release was failing because monit start was installing php and as
it takes a long time to install it, it was problematic for bosh/monit to
acknowledge that simplesaml was running.
Moving the installation of php to a pre start script seems to fix the
problem.

Previous timeout fix was not enough.

Signed-off-by: Camila Londoño <londonoc@vmware.com>